+++ src/test/java/com/arangodb/springframework/testdata/Customer.java
@@ -63,6 +63,9 @@ public class Customer {
 	@Relations(edges = { Owns.class })
 	private Collection<Product> owns;
 
+	@Relations(edges = { Owns.class })
+	private Collection<Product> owns2;
+
 	public Customer() {
 		super();
 	}
@@ -160,6 +163,14 @@ public class Customer {
 
 	public void setOwns(final Collection<Product> owns) {
 		this.owns = owns;
+	}
+
+	public Collection<Product> getOwns2() {
+		return owns2;
+	}
+
+	public void setOwns2(final Collection<Product> owns2) {
+		this.owns2 = owns2;
 	}
 
 	public boolean isAlive() {